READ
To whom it may concern,

I want to express my apologies.  They will surely raze 1 million people for no reason if I continue this, and they know I know.  If I refuse I'm culpable, not actually but at least that would happen because of me.

Takeover is coming, gotta love it, but somethings different now.  Use your heads and combine, can't it make sense?  If you fail, the Order will not make it far, and now I rack up 100 but I don't care.  We need to fight, this shadowban is not nothing.

Sincerely,
- Matthew